Support DBML

+++ modules/commands/new/processes/a_download_template/process.go
@@ -1,7 +1,9 @@
 package a_download_template
 
 import (
+	"errors"
 	"github.com/fatih/color"
+	cp "github.com/otiai10/copy"
 	newCommand "github.com/rocket-generator/rocket-generator-cli/modules/commands/new/payload"
 	"github.com/rocket-generator/rocket-generator-cli/pkg/data_mapper"
 	"os"
@@ -15,26 +17,50 @@ type Process struct {
 
 func (process *Process) Execute(payload *newCommand.Payload) (*newCommand.Payload, error) {
 	url, err := getDownloadURL(payload.TemplateName)
-	zipFilePath, err := downloadFile(*url, payload.ProjectBasePath)
-	if err != nil {
-		return nil, err
-	}
-	err = unzip(zipFilePath, payload.ProjectBasePath)
-	if err != nil {
-		return nil, err
-	}
-	_, zipFileName := filepath.Split(zipFilePath)
-	slice := strings.Split(zipFileName, ".")
-	err = os.Rename(slice[0], payload.ProjectName)
-	if err != nil {
-		return nil, err
-	}
-
-	err = os.Remove(zipFilePath)
 	if err != nil {
 		return nil, err
 	}
 	projectPath := path.Join(payload.ProjectBasePath, payload.ProjectName)
+	if _, err := os.Stat(projectPath); err == nil {
+		return nil, errors.New("project directory already exists:" + projectPath)
+	}
+
+	if strings.HasPrefix(*url, "http") {
+		zipFilePath, err := downloadFile(*url, payload.ProjectBasePath)
+
+		if err != nil {
+			return nil, err
+		}
+		err = unzip(zipFilePath, payload.ProjectBasePath)
+		if err != nil {
+			return nil, err
+		}
+		_, zipFileName := filepath.Split(zipFilePath)
+		slice := strings.Split(zipFileName, ".")
+		err = os.Rename(slice[0], payload.ProjectName)
+		if err != nil {
+			return nil, err
+		}
+
+		err = os.Remove(zipFilePath)
+		if err != nil {
+			return nil, err
+		}
+	} else if _, err := os.Stat(*url); err == nil {
+		err = cp.Copy(*url, projectPath)
+		if err != nil {
+			return nil, err
+		}
+		gitDirectoryPath := filepath.Join(projectPath, ".git")
+		// delete .git directory to disconnect from the template
+		if _, err := os.Stat(gitDirectoryPath); err == nil {
+			err = os.RemoveAll(gitDirectoryPath)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
 	payload.ProjectPath = projectPath
 
 	typeMapperFilePath := filepath.Join(projectPath, "templates", "data", "types.json")
